trigger controller events via transport even if locally transmitted
makes it easier to integrate to smart contract later - we only really need UpdateLocalMetadata to have strong consistency
